requirements.txt
-i https://pypi.org/simple
amqp==5.3.1; python_version >= '3.6'
anyio==4.10.0; python_version >= '3.9'
asgiref==3.9.1; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
autobahn==24.4.2; python_version >= '3.9'
automat==25.4.16; python_version >= '3.9'
billiard==4.2.1; python_version >= '3.7'
celery==5.5.3; python_version >= '3.8'
cffi==2.0.0; platform_python_implementation != 'PyPy'
channels==4.3.1; python_version >= '3.9'
channels-redis==4.3.0; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1.2
click-repl==0.3.0; python_version >= '3.6'
constantly==23.10.4; python_version >= '3.8'
cryptography==45.0.7; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
daphne==4.2.1; python_version >= '3.9'
django==5.2.6; python_version >= '3.10'
django-redis==6.0.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
drf-spectacular==0.28.0; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2; python_version >= '3.8'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.25.1; python_version >= '3.9'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
kombu==5.5.4; python_version >= '3.8'
msgpack==1.1.1; python_version >= '3.8'
packaging==25.0; python_version >= '3.8'
pillow==11.3.0; python_version >= '3.9'
prompt-toolkit==3.0.52; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.23; implementation_name != 'PyPy'
pyjwt==2.10.1; python_version >= '3.9'
pyopenssl==25.1.0
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.1.1
pyyaml==6.0.2; python_version >= '3.8'
redis==6.4.0; python_version >= '3.9'
referencing==0.36.2; python_version >= '3.9'
rpds-py==0.27.1; python_version >= '3.9'
service-identity==24.2.0
setuptools==80.9.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
twisted[tls]==25.5.0; python_full_version >= '3.8.0'
txaio==25.6.1; python_version >= '3.10'
typing-extensions==4.15.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '3.9'
uritemplate==4.2.0; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0
vine==5.1.0; python_version >= '3.6'
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.10.0; python_version >= '3.9'
zope-interface==8.0; python_version >= '3.9'
